+++ R/quill-package.R
@@ -4,7 +4,7 @@
 #' @importFrom rlang enquos eval_tidy quo_is_null quo_get_expr `%||%`
 #' @importFrom vellum vl_scene push pop draw viewport grid_layout unit gpar
 #' @importFrom vellum points_grob lines_grob segments_grob rect_grob text_grob
-#' @importFrom vellum grobwidth grobheight linear_gradient render as_vellum_scene
+#' @importFrom vellum grobwidth grobheight render as_vellum_scene
 NULL
 
 # Names that appear only inside data-masked / after_stat() expressions